Update show.html.erb
Update public profile view to conditionally display group link. It updates the public profile's show view to conditionally display the 'Return to Group Page' link. The link is now wrapped within a conditional statement that checks for the presence of a group. If a group is associated with the student, the link is displayed; otherwise, it is omitted. This change ensures a more accurate and contextual representation of the student's associations, enhancing the user experience and relevance of the page content.